fix: language-specific page IDs are not saved
The translation variable update control is now used. It was previously tried to save the language
variables themselves, problem is, the status when the dev mode is off was not respected and thus the
site ids of matomo were not loaded. it was saved, but not loaded, the second time they were emptied.

fix quiqqer/matomo#11